backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
aiofiles==24.1.0
python-multipart==0.0.20
groq==0.25.0
pydub==0.25.1
imageio-ffmpeg==0.5.1
fpdf2==2.8.3
python-docx==1.1.2
pyinstaller==6.12.0